Belos: MultiVecTraits: Fixed Tpetra spec. for new view semantics.
Tpetra objects, including MultiVector, will change to use view
semantics instead of container semantics.  This means that the
assigment operator (operator=) and copy constructor of MultiVector
will no longer do a deep copy.  This is already true of the "Kokkos
refactor" specializations.

This commit fixes the Tpetra specialization of Belos::MultiVecTraits
to reflect this change.  There is an optimization for CloneCopy() that
prevents a double copy for the old Node types.

NOTE: Anasazi::MultiVecTraits must be fixed in the same way in order
for Anasazi to work correctly with the new Tpetra::MultiVector
specializations.  This commit does NOT attempt to fix Anasazi.  The
necessary changes should be nearly the same as those in this commit.

Build/Test Cases Summary
Enabled Packages: Tpetra, Belos, Ifpack2
0) MPI_DEBUG => passed: passed=141,notpassed=0 (7.10 min)
1) SERIAL_RELEASE => passed: passed=121,notpassed=0 (8.86 min)
Other local commits for this build/test group: d22a8a9, 99faf1c