Ifpack: sped up Chebyshev ApplyInverse() and minor readibility changes
Chebyshev was copying the rhs vector to three additional vectors (V, W,
and Temp) during each ApplyInverse. By looking at the code, one notices
that those vectors are rewritten before their first use. As such, we
only need to allocate the memory for them, but do not need to fill in
during construction.

Ideally, those vectors would have been moved to the Compute() call, but
unfortunately we don't know how many vectors are in a multivector at
that point.

The patch gave 2x speed up on a 100^3 3D Laplace equation.