Ifpack2: ReorderFilter: Removed dependence on Zoltan2.
ReorderFilter's constructor was taking a Zoltan2 object, but not
keeping it.  It just extracted the permutation and inverse permutation
arrays from it as ArrayRCP.  This means that ReorderFilter doesn't
need to depend on Zoltan2, so I removed this dependency by changing
the constructor so that it takes the two permutation arrays directly.
AdditiveSchwarz still depends on Zoltan2 because it invokes Zoltan2 to
compute the permutation.

ReorderFilter already had a unit test which creates the ReorderFilter
instance in the same way that AdditiveSchwarz does.  The unit test
referred to Zoltan2, but doesn't actually use its reordering; it
computes an artificial reordering, even though it asks Zoltan2 for
arrays in which to put the reordering.  I changed the test so that it
still makes a ReorderFilter even if Ifpack2 was built without Zoltan2.
It never needed to depend on Zoltan2, so I removed this dependency and
the associated includes of Zoltan2 classes.