KokkosCompat: there was still an issue with static nodes. Hopefully fixed
Added a check if HostSpace was already destroyed to the destructors of the nodes.
That way only a warning is printed that the device was not terminated in time,
without causing segfaults when actually terminating the device.

Fixed a small issue in MultiVector, with calling an unsuported update function.

ToDo:
Have to make sure that MV_add works if the type of scalar prefactors is
different from the Multivector scalars.