MueLu: add barrier to Repartitition test
This commit inserts a barrier between the first and second hierarchy
construction, just after the first one is destroyed.

It appears the CoordinateMap test is hitting a parallel race condition.  It
would be surprising if replacing the coordinate vector map were causing this.
I've captured the stack, but don't have time to look at this right now.