Stokhos:  "Fix" really annoying warning about volatile return values.
I get tons of warnings along the lines of:

"warning: implicit dereference will not access object of type ..."

having to do with the return type of a volatile function (e.g., the
assignment operator) not being used by the calling code.  Apparently GCC
does this to tell you the volatile object will not be read, which is
exactly what you would expect anyway.

I decided to "fix" this by casting away volatileness on the return type
for those functions, which I think in the volatile world should be safe,
e.g.,

a = b = c

for volatile b should still work (although I haven't tested this).