Kokkos: Fixed build warnings for declared but unused typedefs.
In some cases, those typedefs were being used in compile-time
assertions, for example in asserting that a particular shape is rank
one.  I did not remove those typedefs; instead, where possible, I
declared an empty or trivial instance of that type in an inner scope,
and protected it with "(void)" so it would not cause an "unused
variable" warning.  Hopefully this does not introduce an excessive
run-time cost.