Thyra: Adding adjoint support for row sum and scaling of sub-operators in the blocked linear operator
Also added a test that excercises a matrix of the form

   [A  u]
   [v' 0]

where A is a square epetra operator and u and v are multi-vectors
with two columns. This supports a specific use case for computing
the row sum and scaling the matrix by that.

Also fixing a few test issues from the previous commits.