Tpetra: RTI: Added output to help debug RTI CG test.
RTI's CG test fails in SERIAL_RELEASE.  It looks like CG runs, but zeros
out everything (including the current solution vector and the internal
residual).  This makes CG "converge," but it computes a solution vector
of zero, which is wrong.

I'm a bit worried that the changes I made recently (e.g., added Pair
with volatile overloads) broke something, though I think that code is
only used by a different example, not in RTI core or by this example.

The test still fails, but the output added in this commit will help with
debugging.