Tpetra: CrsMatrix (not Kokkos refactor yet): Fixed convert.
Thanks to Eric Phipps for pointing out that my recent commit to
Tpetra_CrsMatrix_def.hpp that optimized convert() broke that method.
(The commit only affected the "classic" CrsMatrix, not the Kokkos
refactor version.) In fact, convert() should have returned null
unconditionally.  Does Tpetra have a test for convert()?

This commit fixes the issue.  Note that the Kokkos refactor version of
CrsMatrix::convert() did not have the optimization of the commit in
question, and thus did not share the bug.  I am reluctant to port the
optimization to the Kokkos refactor version until there is a test for
convert().  It should not be hard to write one.

Build/Test Cases Summary
Enabled Packages: Tpetra
0) MPI_DEBUG => passed: passed=72,notpassed=0 (1.63 min)
1) SERIAL_RELEASE => passed: passed=56,notpassed=0 (0.29 min)