MueLu: reenabling PgPFactory test
It still throws, but the reason is known: incorrect matrix
redistribution in Amesos (see bug #969). As it is an important bug for
us that we need to address and not ignore, I'm reenabling this test to
be a visible remindment of that.

Problem summary: during import of a matrix to a single processor, Amesos
Superlu assumes that GIDs are from 0 to the max, which may not be the
case. Some other Amesos solvers were fixed, but not SuperLU, which is
unfortunate, as that is the main working horse for us.