Kokkos,Tpetra:  Introduce InnerProductTraits into KokkosRefactor.
Setup Tpetra Vector/MultiVector in the KokkosRefactor branch to use
InnerProductSpaceTraits to define the return type of dot() as well as
use the scalar-level dot() function in computing the dot product.  This
is important for embedded UQ scalar types that redefine what this means
instead of just using '*'.  For compatibility, I also added the dot_type
typedef in the original Vector/MultiVector branch, but is is just equal
to the scalar type in that case.