Stokhos:  Incorporate Kokkos MpiComm and InnerProductSpaceTraits.
Added overloads of KokkosMpiComm functions for Sacado::MP::Vector and
implemented InnerProductSpaceTraits for this scalar type.  I also
changed the implementation in ArithTraits to compute the proper
magnitude.  This gets the Tpetra stack with the MP::Vector scalar type
working for the dynamically sized storage type, and all of those tests
now pass in parallel.

Included in this commit are also a number of fixes for compiler
errors/warnings for compiling the stack with NVCC.