Automatically selecting default rst2thml[.py] and rst2latex[.py]
I put in a simple test that I used for 'eg' in CheckinTest.py to see if we
should pick 'rst2thml' or 'rs2thml.py' as the default for --generate-html.
This will make it so that I don't have to keep manually switching these
arguments on the command-line between the Fissile machines and my laptop.