Tpetra: CrsGraph (Kokkos refactor): Fixed build warning in new ctor.
The new constructor in question takes a row Map, a column Map, a local
(new Kokkos) graph, and an optional Teuchos::ParameterList.  This
constructor should and does only exist in the Kokkos refactor version
of CrsGraph.  I fixed a build warning related to the constructor's
initialization list initializing a field in the wrong order.  I also
took the chance to improve run-time error checks and code formatting.