Tpetra: CrsMatrix (classic & KR): "Preserve Local Graph" defaults true.
In CrsMatrix, in both the classic and Kokkos refactor versions, the
undocumented parameter "Preserve Local Graph" now defaults to true.
This makes the following scenario work by default:

  1. Create a CrsMatrix A that creates and owns its graph (i.e., don't
     use the constructor that takes an RCP<const Tpetra::CrsGraph> or
     a local graph)
  2. Set an entry in the matrix A, and call fillComplete on it
  3. Create a CrsMatrix B using A's graph (obtained via
     A.getCrsGraph()), so that B has a const (a.k.a. "static") graph
  4. Change a value in B (you can't change its structure), and call
     fillComplete on B

Before this commit, the above scenario didn't work by default.  This
is because A's first fillComplete call would call
fillLocalGraphAndMatrix, which by default sets the local graph to
null.  As a result, from that point, A.getCrsGraph()->getLocalGraph()
returns null, which makes B's fillComplete throw an exception.  The
only way to make this scenario work was to set A's "Preserve Local
Graph" parameter to true.  (It defaulted to false.)

The idea behind this nonintuitive behavior was for the local sparse
ops object to own all the data.  This might make sense if it is a
third-party library that takes CSR's three arrays and copies them into
its own storage format.  In that case, it might be a good idea to free
the original three CSR arrays, in order to avoid duplicate storage.
However, resumeFill never had a way to get that data back out of the
local sparse ops object.  Rather than try to implement that, it's
easier just to make "Preserve Local Graph" default to true.

The possible data duplication mentioned in the previous paragraph can
never happen with the Kokkos refactor version of CrsMatrix, since it
insists on controlling the matrix representation itself.  This makes
the code shorter and easier to read, and also ensures efficient fill.
That will in turn make the option unnecessary.

Thanks to Andrey for pointing this out and giving a test case.

Build/Test Cases Summary
Enabled Packages: KokkosCore, KokkosCompat, KokkosLinAlg, KokkosMpiComm, KokkosContainers, Tpetra
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_KOKKOS_REFACTOR => passed: passed=74,notpassed=0 (4.38 min)
3) SERIAL_RELEASE_KOKKOS_REFACTOR => passed: passed=58,notpassed=0 (6.78 min)