Tpetra: Fixed bugs in the new constructors taking Kokkos objects
The new constructors in CrsGraph and CrsMatrix were throwing exceptions in Debug mode because
the crs arrays which are directly part of the Tpetra classes were not created. Those arrays are
typically not used if the Tpetra objects are fillComplete (which is the case for those constructors)
but their existence is checked for in Debug mode. Note: the rowPtrs array of CrsGraph can not be created:
the Kokkos::StaticCrsGraph has a const row_map array (View<const inttype* ...>) and thus can not be used
without some hacks to create the non const rowPtrs thing in Tpetra::CrsGraph.