Changing name of CTestCustom.ctest to CTestCustom.cmake and added trace
According to the article:

   http://www.kitware.com/blog/home/post/27

a project should be using CTestCustom.cmake and not CTestCustom.ctest to avoid
expensive recursive searches for subordinate CTestCustom.ctest files in the
binary tree.

This breaks backward compatibility but I think Trilinos is the only project
that used this file and I can change that.

This is another file that gets magically processed by TriBITS if it exists so
I added a trace command and added the file trace category 'CONFIGURE'.

I added this to TribitsExampleProject and updated the test build to show that
the file is being processed.