Tpetra: Map (classic only): Removed dead code.
Map (not the Kokkos refactor version) used to have two compile-time
options for its GID->LID lookup table: a dynamic-structure hash table,
and a fixed-structure hash table.  The latter is always a better idea,
so right after we implemented it last spring, it became the default.
This commit eliminates the dynamic-structure compile-time option,
since it's dead code.

Note that this distinction does not apply to the Kokkos refactor
version of Map.  That only uses Kokkos::UnorderedMap (currently) for
the GID->LID lookup table.  UnorderedMap is (sort of) a
dynamic-structure hash table.  Thus, it may still pay to use an even
more restrictive fixed-structure table, like in the
non-Kokkos-refactor version of Map.  However, the payoff may be less,
since the original dynamic-structure hash table does dynamic memory
allocation on each insert, while Kokkos::UnorderedMap does not.