Tpetra: MV (Kokkos ref): Eliminated ArrayRCP "views" used by DistObject.
Tpetra "classic" assumed that MPI could only read from and write to
host memory, not CUDA device memory.  (This was true in 2008, but is
no longer true; MPI implementations now can access CUDA device memory
directly.)  As a result, Tpetra::DistObject had hooks for telling the
subclass to create and relinquish host views of the subclass' data.

The Kokkos refactor version of Tpetra::MultiVector does not need this
facility, since it assumes that MPI can read CUDA device memory.
Thus, I removed two Teuchos::ArrayRCP that stored these views (two
ArrayRCP: one for const, for the source object, and the other for
nonconst, for the destination object) and the code that created and
relinquished them.