Splitting package/subpackages enable/disable, added another example
With this commit I fixed a nasty little bug where the enable of the parent
package would be performed before the disables were applied.  This would
result in an enable that had to be overridden and if
${PROJECT_NAME}_DISABLE_ENABLED_FORWARD_DEP_PACKAGES=OFF, this would even
cause the configure to fail!  Andy Salinger had mentioned this but I can't
find a Trilinos ticket for this.

To fix this I split up the package/subpackages enable and disable loops
appropriately.  Now all the disables should be applied correctly first any
enables so an enable override will not occur unless there is a true conflicit.

I added a ReducedMockTrilinos example/test that showed this bad behavior and I
fixed it.  There was another test that I change to change to show the new
behvaior.