Tpetra: Vector (Kokkos refactor): Impl'd offsetView(NonConst).
offsetView and offsetViewNonConst were not implemented in the Kokkos
refactor version of Tpetra::Vector.  This is because in the "classic"
implementations of these methods, the necessary "expert view
constructor" was missing.  It turns out that the Kokkos refactor version
of Vector already has a constructor that takes a Map and a DualView, and
these are the constructors we want.  Fixing this will fix build failures
on perseus.  The issue only shows up in Ifpack2::LocalFilter, which uses
these methods.

There is a possible bug in both the classic and Kokkos refactor versions
of these methods, relating to whichVectors_.  Please see the internal
comments in the Kokkos refactor versions.

Build/Test Cases Summary
Enabled Packages: Tpetra
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=72,notpassed=0 (1.63 min)
1) SERIAL_RELEASE => passed: passed=56,notpassed=0 (0.29 min)