Teuchos: Moved toString(EReductionType) implementation to .cpp file.
The implementation of toString(EReductionType) (the overload of
toString() for an EReductionType enum input argument) didn't need to be
in the header file, but was, and was even marked inline.  It's not
performance critical, so there was no need to mark it inline or force
users to compile it over and over again.  Furthermore, the
implementation has a switch statement which results in the classic
dilemma [1] about compiler warnings for unreachable statements.  Since
the implementation lived in the header file, all source files that
included Teuchos_CommHelpers.hpp (e.g., those of most Tpetra classes)
were exposed to a potential compiler warning.  I fixed this by moving
the implementation from the header file to the source (.cpp) file.

1. The switch's 'default' clause throws an exception.  Some compilers
warn (unnecessarily) in this case that the switch statement has a branch
which does not return.  Attempts to fix this by inserting a superfluous
return statement after the exception throw cause _other_ compilers to
warn (legitimately) about the unreachable statement.  Thus, it's
impossible to please all compilers.

Build/Test Cases Summary
Enabled Packages: TeuchosComm
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=72,notpassed=0 (1.63 min)
1) SERIAL_RELEASE => passed: passed=56,notpassed=0 (0.29 min)