Tpetra: Directory now knows whether it is "locally one-to-one."
It's easy for Directory to know whether any of the (PID,LID) pairs it
owns map to duplicate GIDs.  If this is true on any process in the
Directory's communicator, then the Directory's input Map is _not_ one to
one; otherwise, the input Map is one to one.  We can use this to add an
inexpensive (at most one all-reduce) test for whether a Map is one to
one.  This commit adds the "is locally one to one?" predicate to
Directory.

The reason we care about this, is because several outstanding Ifpack2
and Tpetra issues relate to whether the input Map is one to one.  Thus,
it's helpful to have an inexpensive test for this.