Fixing to work on Windows with MSVS genereator
I had to disable a few tests on Windows involving OUTPUT_FILE.  It seems that
EXECUTE_PROCESS() does not close an ouptut file properly before it gets read.
Very irritating.

These tests only work if the commit:

  'Fixing TRIBITS_ADD_TEST() for MSVS generator (PHI Kanban #3314)'

is included.  If you revert that commit and run these tests on Windows with
MSVS generator a bunch of them fail!