Added something for major todo, added more extract function/macro doc
With this commit, I have addressed every major plain-text todo in the entire
main document.  Now every section I had outlined has at last something in that
section. There are still a lot of todos in comments.  Also, there are still
plain-text todos in some of the extracted function/macro RST blocks that I
need to addresss.

Now I just need to read through and spell check this entire mess and then I
can post this as the initial version of this document and ask for some
feedback.