Kokkos: Fix configure error when secondary stable code is off
Turning KokkosCompat to secondary stable changed the behavior of
Trilinos_ENABLE_KokkosCompat to be blank instead of OFF. This exposed an issue
in the KokkosClassic cmake files where that variable being black caused a macro
to have too few parameters. Quoting the variable fixed this so that even in the
blank case there is something for a parameter, but it still evaluates as
expected for logical tests.