Update Perseus testing again
The Kokkos subpackages needed to be set back to experimental temporarily due to
an issue with the gcc 4.1.2 compilers. I am explicitly enabling the Kokkos
subpackages here so that they are still enabled. This method is dangerous as it
enables them for all package builds, but is necessary since these subpackages
are experimental now. The risk of causing spam associated with this should be
minimal for the short time that this work around is needed.