Ifpack2: Added "partitioner: map" to the Ifpack2 parameter list.
This commit actually belongs with my previous commit. If the user
selects "user" as the PartitionerType for BlockRelaxation, then a
"map" is required. (This is perhaps a bad word choice here, but I
copied Ifpack.) The "map" assigns each local row in the matrix
(whatever matrix eventually makes it to BlockRelaxation, that is) a
"partition ID". This is the nonoverlapping partition to which that row
belongs.