Ifpack2: Fixed a bug in DenseContainer
Upon doing a test (to be committed soon), I noticed that there was a
problem with BlockRelaxation. Namely, it wasn't giving the right
answer. I managed to trace the error back to DenseContainer. There was
some nonsense going on with its "ifndef" preprocessor stuff, and the
def file was trying to include itself. Beyond that, some funny things
were happening in the applyImpl(). These have all been fixed and
BlockRelaxation now gives the correct answer.