Tpetra: CrsMatrix: fillComplete w/ const graph now uses graph's Maps.
The zero- and one-argument versions of fillComplete pick default
domain and range Maps for the user.  This is fine.  However, for some
reason lost to the mists of time, fillComplete always uses the row Map
as the domain and range Maps, even if the matrix has a fill-complete
constant graph.  In that case, the graph already has domain and range
Maps, which could very well differ from the row Map.  That causes
trouble for users, who have to remember when calling the matrix's
fillComplete to provide the domain and range Maps, even though they
have already given them to the graph's fillComplete.

This commit tests whether the matrix has a constant ("static") graph.
If it does, the zero- and one-argument versions of fillComplete always
use the constant graph's domain resp. range Maps as the default domain
resp. range Maps of the matrix.  (The two- and three-argument versions
of fillComplete, which take domain and range Maps, do not currently do
this.  We assume that if users pass in domain and range Maps
explicitly, then they know what they are doing.)

This commit also fixes an issue with getAllValues() in the Kokkos
Refactor version of CrsMatrix.  Namely, that method was always using
myGraph_ (the nonconst graph), when it should use getCrsGraph() (which
returns whichever of the const or nonconst graphs is nonnull, in that
order).  I fixed this and also added more error checking.

Thanks to Christian for pointing out this issue to me.