Xpetra: TpetraCrsMatrix: Use existing Node instead of creating new one.
TpetraCrsMatrix's copy constructor (which does a deep copy via
Tpetra::CrsMatrix::clone()) was creating a new Kokkos Node instance,
instead of using the input TpetraCrsMatrix's Node instance.  That's
not incorrect, but it's not necessary.  This commit fixes that.

This has nothing to do with the failing Xpetra test mentioned in my
previous commit; I just wanted to fix it because I saw it.