Seacas: more cmake file updates
Updated the exodus cmake files to glob all the source files instead of manually
specify them. This should prevent the need to update the cmake build when new
files are added.

Also had to temporarily disable strong warning/error flags for all of Seacas.
The reason is due to the usage of some c99 features in exodus_for and the fact
that currently the strong warnings set ansi which enforces c89. There has been a
proposal made in Trilinos to change the strong warnings to c99 however, that
won't be resolved this week for sure so putting in a temporary fix to avoid
having this prevent integration updates from Sierra.

(From Sierra)