ML: fixing compile error if Teuchos disabled
In Bug#6116, Karen Devine reported that the ML scaling test was not
compiling if Teuchos was disabled.  This was due to the fact that that
said test uses the Teuchos CommandLineProcessor, but there were no guards
in the CMakeLists.txt file to check whether Teuchos was enabled.

I was able to reproduce the error and verify that this commit fixes it.