Tpetra: CrsMatrix (classic): Doc'd & improved transferAndFillComplete.
The transferAndFillComplete method is a clever innovation by Chris
Siefert to improve Tpetra's sparse matrix-matrix multiply performance.
He provided an implementation recently for the "classic" version of
CrsMatrix.  In preparation for porting this to the Kokkos refactor
version of CrsMatrix, I added lots of internal documentation
explaining potential porting pitfalls.  I also moved error checking
closer to the beginning of the method, in order to catch errors
faster.  Finally, I improved the public documentation of this method,
importAndFillComplete, and exportAndFillComplete, and checked how
these methods are used in RowMatrixTransposer.