Changing SEND_ERROR to FATAL_ERROR for extra repo processing errors
Before, if there was an error in processing the list of extra repos, it would
just do MESSAGE(SEND_ERROR ...) but continue the configure.  When the
configure takes 10 minutes (like it currently does with CASL VERA), the error
is shown right away but the configure does not stop until 10 minutes later.

For errors that occur very early in the configure (like extra repo processing
errors do), it makes sense to fail fast with FATAL_ERROR.  However, for errors
that occur late in configure, it makes sense to do SEND_ERROR so that you can
see several errors before processing stops.  For example, most compilers don't
stop compilation on the first error, they report all errors.

I had to fix a few unit tests to ignore the missing Dakota repo with
MockTrilinos.  These are better tests now.  It was bad for these configures to
be producing configure errors but still passing due to the right regexes at
the end.