TrilinosCoupling/FENL:  Add Belos/MueLu solve.
Added Belos and MueLu to FENL.  This can be used instead of the native
CG solver that was already there.  Add "belos" to the command line to
get the Belos solver, and "muelu" to turn on MueLu preconditioning.
Currently it is hard-coded to read muelu.xml for the MueLu settings.
Note, there is currently a problem with the RAP in MueLu for
KokkosRefactor that this xml file exposes.

Thanks to Jonathan for most of this work.